+++ src/adapter/src/coord/read_policy.rs
@@ -147,7 +147,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
                 if let Some(timeline) = self.get_timeline(*id) {
                     let initial_frontier = self
                         .controller
-                        .compute(*compute_instance)
+                        .compute
+                        .instance(*compute_instance)
                         .unwrap()
                         .collection(*id)
                         .unwrap()
@@ -179,7 +180,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
                 compute_policy_updates.push((*id, self.read_capability[&id].policy()));
             }
             self.controller
-                .active_compute(*compute_instance)
+                .active_compute()
+                .instance(*compute_instance)
                 .unwrap()
                 .set_read_policy(compute_policy_updates)
                 .await
@@ -198,7 +200,7 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             if let Some(timeline) = self.get_timeline(*id) {
                 let initial_frontier = self
                     .controller
-                    .storage()
+                    .storage
                     .collection(*id)
                     .unwrap()
                     .read_capabilities
@@ -223,7 +225,7 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             storage_policy_updates.push((*id, self.read_capability[&id].policy()));
         }
         self.controller
-            .storage_mut()
+            .storage
             .set_read_policy(storage_policy_updates)
             .await
             .unwrap();
@@ -241,7 +243,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             .expect("coord out of sync");
         capability.base_policy = base_policy;
         self.controller
-            .active_compute(compute_instance)
+            .active_compute()
+            .instance(compute_instance)
             .unwrap()
             .set_read_policy(vec![(id, capability.policy())])
             .await
@@ -262,9 +265,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
     pub(crate) async fn acquire_read_holds(&mut self, read_holds: &ReadHolds<mz_repr::Timestamp>) {
         // Update STORAGE read policies.
         let mut policy_changes = Vec::new();
-        let storage = self.controller.storage_mut();
         for id in read_holds.id_bundle.storage_ids.iter() {
-            let collection = storage.collection(*id).unwrap();
+            let collection = self.controller.storage.collection(*id).unwrap();
             assert!(
                 collection
                     .read_capabilities
@@ -279,13 +281,18 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             read_needs.holds.update_iter(Some((read_holds.time, 1)));
             policy_changes.push((*id, read_needs.policy()));
         }
-        storage.set_read_policy(policy_changes).await.unwrap();
+        self.controller
+            .storage
+            .set_read_policy(policy_changes)
+            .await
+            .unwrap();
         // Update COMPUTE read policies
         for (compute_instance, compute_ids) in read_holds.id_bundle.compute_ids.iter() {
             let mut policy_changes = Vec::new();
-            let mut compute = self.controller.active_compute(*compute_instance).unwrap();
+            let mut compute = self.controller.active_compute();
+            let mut instance = compute.instance(*compute_instance).unwrap();
             for id in compute_ids.iter() {
-                let collection = compute.collection(*id).unwrap();
+                let collection = instance.collection(*id).unwrap();
                 assert!(collection
                     .read_frontier()
                     .less_equal(&read_holds.time),
@@ -299,7 +306,7 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
                 read_needs.holds.update_iter(Some((read_holds.time, 1)));
                 policy_changes.push((*id, read_needs.policy()));
             }
-            compute.set_read_policy(policy_changes).await.unwrap();
+            instance.set_read_policy(policy_changes).await.unwrap();
         }
     }
     /// Update the timestamp of the read holds on the indicated collections from the
@@ -324,9 +331,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
 
         // Update STORAGE read policies.
         let mut policy_changes = Vec::new();
-        let storage = self.controller.storage_mut();
         for id in storage_ids.iter() {
-            let collection = storage.collection(*id).unwrap();
+            let collection = self.controller.storage.collection(*id).unwrap();
             assert!(collection
                 .read_capabilities
                 .frontier()
@@ -342,13 +348,18 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             read_needs.holds.update_iter(Some((*old_time, -1)));
             policy_changes.push((*id, read_needs.policy()));
         }
-        storage.set_read_policy(policy_changes).await.unwrap();
+        self.controller
+            .storage
+            .set_read_policy(policy_changes)
+            .await
+            .unwrap();
         // Update COMPUTE read policies
+        let mut compute = self.controller.active_compute();
         for (compute_instance, compute_ids) in compute_ids.iter() {
             let mut policy_changes = Vec::new();
-            let mut compute = self.controller.active_compute(*compute_instance).unwrap();
+            let mut instance = compute.instance(*compute_instance).unwrap();
             for id in compute_ids.iter() {
-                let collection = compute.collection(*id).unwrap();
+                let collection = instance.collection(*id).unwrap();
                 assert!(collection
                     .read_frontier()
                     .less_equal(&new_time),
@@ -364,7 +375,7 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
                 read_needs.holds.update_iter(Some((*old_time, -1)));
                 policy_changes.push((*id, read_needs.policy()));
             }
-            compute.set_read_policy(policy_changes).await.unwrap();
+            instance.set_read_policy(policy_changes).await.unwrap();
         }
 
         read_holds.time = new_time;
@@ -395,11 +406,12 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
             }
         }
         self.controller
-            .storage_mut()
+            .storage
             .set_read_policy(policy_changes)
             .await
             .unwrap();
         // Update COMPUTE read policies
+        let mut compute = self.controller.active_compute();
         for (compute_instance, compute_ids) in compute_ids.iter() {
             let mut policy_changes = Vec::new();
             for id in compute_ids.iter() {
@@ -409,8 +421,8 @@ impl<S: Append + 'static> crate::coord::Coordinator<S> {
                     policy_changes.push((*id, read_needs.policy()));
                 }
             }
-            if let Some(mut compute) = self.controller.active_compute(*compute_instance) {
-                compute.set_read_policy(policy_changes).await.unwrap();
+            if let Some(mut instance) = compute.instance(*compute_instance) {
+                instance.set_read_policy(policy_changes).await.unwrap();
             }
         }
     }
