+++ src/controller/Cargo.toml
@@ -10,8 +10,6 @@ publish = false
 anyhow = "1.0.64"
 chrono = { version = "0.4.22", default-features = false, features = ["clock", "std"] }
 differential-dataflow = { git = "https://github.com/TimelyDataflow/differential-dataflow.git" }
-futures = "0.3.24"
-maplit = "1.0.2"
 mz-build-info = { path = "../build-info" }
 mz-compute-client = { path = "../compute-client" }
 mz-orchestrator = { path = "../orchestrator" }
@@ -21,11 +19,7 @@ mz-proto = { path = "../proto" }
 mz-ore = { path = "../ore" }
 mz-repr = { path = "../repr" }
 mz-storage = { path = "../storage" }
-once_cell = "1.14.0"
-prost = { version = "0.11.0", features = ["no-recursion-limit"] }
-regex = "1.6.0"
 serde = { version = "1.0.144", features = ["derive"] }
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", default-features = false, features = ["bincode"] }
 tokio = "1.19.2"
-tracing = "0.1.36"
 uuid = { version = "1.1.2" }
