+++ src/controller/src/lib.rs
@@ -21,19 +21,11 @@
 //! Consult the `StorageController` and `ComputeController` documentation for more information
 //! about each of these interfaces.
 
-use std::collections::{BTreeMap, BTreeSet};
 use std::mem;
-use std::num::NonZeroUsize;
 use std::sync::Arc;
 
-use anyhow::anyhow;
 use chrono::{DateTime, Utc};
 use differential_dataflow::lattice::Lattice;
-use futures::future::{self, FutureExt};
-use futures::stream::{BoxStream, StreamExt};
-use maplit::hashmap;
-use once_cell::sync::Lazy;
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 use timely::order::TotalOrder;
 use timely::progress::Timestamp;
@@ -41,19 +33,13 @@ use tokio::sync::Mutex;
 use uuid::Uuid;
 
 use mz_build_info::BuildInfo;
-use mz_compute_client::command::{ComputeCommand, ProcessId, ProtoComputeCommand, ReplicaId};
+use mz_compute_client::command::ReplicaId;
 use mz_compute_client::controller::{
     ActiveComputeController, ComputeController, ComputeControllerResponse, ComputeInstanceId,
 };
-use mz_compute_client::logging::{LogVariant, LogView, LoggingConfig};
-use mz_compute_client::response::{
-    ComputeResponse, PeekResponse, ProtoComputeResponse, TailResponse,
-};
+use mz_compute_client::response::{PeekResponse, TailResponse};
 use mz_compute_client::service::{ComputeClient, ComputeGrpcClient};
-use mz_orchestrator::{
-    CpuLimit, LabelSelectionLogic, LabelSelector, MemoryLimit, NamespacedOrchestrator,
-    Orchestrator, ServiceConfig, ServiceEvent, ServicePort,
-};
+use mz_orchestrator::Orchestrator;
 use mz_ore::tracing::OpenTelemetryContext;
 use mz_persist_client::cache::PersistClientCache;
 use mz_persist_client::PersistLocation;
@@ -64,8 +50,6 @@ use mz_storage::controller::StorageController;
 use mz_storage::protocol::client::{
     ProtoStorageCommand, ProtoStorageResponse, StorageCommand, StorageResponse,
 };
-
-pub use mz_orchestrator::ServiceStatus as ComputeInstanceStatus;
 
 /// Configures a controller.
 #[derive(Debug, Clone)]
@@ -86,138 +70,6 @@ pub struct ControllerConfig {
     pub storaged_image: String,
     /// The computed image to use when starting new compute processes.
     pub computed_image: String,
-}
-
-/// Resource allocations for a replica of a compute instance.
-#[derive(Copy, Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
-pub struct ComputeInstanceReplicaAllocation {
-    /// The memory limit for each process in the replica.
-    pub memory_limit: Option<MemoryLimit>,
-    /// The CPU limit for each process in the replica.
-    pub cpu_limit: Option<CpuLimit>,
-    /// The number of processes in the replica.
-    pub scale: NonZeroUsize,
-    /// The number of worker threads in the replica.
-    pub workers: NonZeroUsize,
-}
-
-/// Size or location of a replica
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum ConcreteComputeInstanceReplicaLocation {
-    /// Out-of-process replica
-    Remote {
-        /// The network addresses of the processes in the replica.
-        addrs: BTreeSet<String>,
-    },
-    /// A remote but managed replica
-    Managed {
-        /// The resource allocation for the replica.
-        allocation: ComputeInstanceReplicaAllocation,
-        /// SQL size parameter used for allocation
-        size: String,
-        /// The replica's availability zone
-        availability_zone: String,
-        /// `true` if the AZ was specified by the user and must be respected;
-        /// `false` if it was picked arbitrarily by Materialize.
-        az_user_specified: bool,
-    },
-}
-
-impl ConcreteComputeInstanceReplicaLocation {
-    pub fn get_az(&self) -> Option<&str> {
-        match self {
-            ConcreteComputeInstanceReplicaLocation::Remote { .. } => None,
-            ConcreteComputeInstanceReplicaLocation::Managed {
-                availability_zone, ..
-            } => Some(&*availability_zone),
-        }
-    }
-}
-
-/// Logging configuration of a replica.
-/// Changing this type requires a catalog storage migration!
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-pub enum ConcreteComputeInstanceReplicaLogging {
-    /// Instantiate default logging configuration upon system start.
-    /// To configure a replica without logging, ConcreteViews(vec![],vec![]) should be used.
-    Default,
-    /// Logging sources and views have been built for this replica.
-    ConcreteViews(Vec<(LogVariant, GlobalId)>, Vec<(LogView, GlobalId)>),
-}
-
-impl ConcreteComputeInstanceReplicaLogging {
-    /// Return all persisted introspection sources contained
-    pub fn get_sources(&self) -> Vec<(LogVariant, GlobalId)> {
-        match self {
-            ConcreteComputeInstanceReplicaLogging::Default => vec![],
-            ConcreteComputeInstanceReplicaLogging::ConcreteViews(logs, _) => logs.clone(),
-        }
-    }
-
-    /// Return all persisted introspection views contained
-    pub fn get_views(&self) -> Vec<(LogView, GlobalId)> {
-        match self {
-            ConcreteComputeInstanceReplicaLogging::Default => vec![],
-            ConcreteComputeInstanceReplicaLogging::ConcreteViews(_, views) => views.clone(),
-        }
-    }
-
-    /// Return all ids of the persisted introspection views contained
-    pub fn get_view_ids(&self) -> Vec<GlobalId> {
-        self.get_views().into_iter().map(|(_, id)| id).collect()
-    }
-
-    /// Return all ids of the persisted introspection sources contained
-    pub fn get_source_ids(&self) -> Vec<GlobalId> {
-        self.get_sources().into_iter().map(|(_, id)| id).collect()
-    }
-
-    /// Return all ids of the persisted introspection sources and logs contained
-    pub fn get_source_and_view_ids(&self) -> Vec<GlobalId> {
-        self.get_source_ids()
-            .into_iter()
-            .chain(self.get_view_ids().into_iter())
-            .collect()
-    }
-}
-
-/// Replica configuration
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct ConcreteComputeInstanceReplicaConfig {
-    pub location: ConcreteComputeInstanceReplicaLocation,
-    pub persisted_logs: ConcreteComputeInstanceReplicaLogging,
-}
-
-/// Deterministically generates replica names based on inputs.
-fn generate_replica_service_name(instance_id: ComputeInstanceId, replica_id: ReplicaId) -> String {
-    format!("cluster-{instance_id}-replica-{replica_id}")
-}
-
-/// Parse a name generated by `generate_replica_service_name`, to extract the
-/// replica's compute instance ID and replica ID values.
-fn parse_replica_service_name(
-    service_name: &str,
-) -> Result<(ComputeInstanceId, ReplicaId), anyhow::Error> {
-    static SERVICE_NAME_RE: Lazy<Regex> =
-        Lazy::new(|| Regex::new(r"(?-u)^cluster-(\d+)-replica-(\d+)$").unwrap());
-
-    let caps = SERVICE_NAME_RE
-        .captures(service_name)
-        .ok_or_else(|| anyhow!("invalid service name: {service_name}"))?;
-
-    let instance_id = caps.get(1).unwrap().as_str().parse().unwrap();
-    let replica_id = caps.get(2).unwrap().as_str().parse().unwrap();
-    Ok((instance_id, replica_id))
-}
-
-/// An event describing a change in status of a compute process.
-#[derive(Debug, Clone, Serialize)]
-pub struct ComputeInstanceEvent {
-    pub instance_id: ComputeInstanceId,
-    pub replica_id: ReplicaId,
-    pub process_id: ProcessId,
-    pub status: ComputeInstanceStatus,
-    pub time: DateTime<Utc>,
 }
 
 /// Responses that [`Controller`] can produce.
@@ -267,279 +119,20 @@ enum Readiness {
 
 /// A client that maintains soft state and validates commands, in addition to forwarding them.
 pub struct Controller<T = mz_repr::Timestamp> {
-    build_info: &'static BuildInfo,
-    storage_controller: Box<dyn StorageController<Timestamp = T>>,
-    compute_orchestrator: Arc<dyn NamespacedOrchestrator>,
-    computed_image: String,
-    compute: BTreeMap<ComputeInstanceId, ComputeController<T>>,
+    pub storage: Box<dyn StorageController<Timestamp = T>>,
+    pub compute: ComputeController<T>,
     readiness: Readiness,
-    /// Set to `true` once `initialization_complete` has been called.
-    initialized: bool,
-}
-
-impl<T> Controller<T>
-where
-    T: Timestamp + Lattice + Codec64 + Copy + Unpin,
-    ComputeCommand<T>: RustType<ProtoComputeCommand>,
-    ComputeResponse<T>: RustType<ProtoComputeResponse>,
-{
-    pub async fn create_instance(
-        &mut self,
-        instance: ComputeInstanceId,
-        logging: Option<LoggingConfig>,
-        max_result_size: u32,
-    ) {
-        // Insert a new compute instance controller.
-        self.compute.insert(
-            instance,
-            ComputeController::new(instance, self.build_info, &logging, max_result_size),
-        );
-        if self.initialized {
-            self.compute_mut(instance)
-                .expect("Instance just initialized")
-                .initialization_complete();
-        }
-    }
-
-    /// Adds replicas of an instance.
-    ///
-    /// # Panics
-    /// - If the identified `instance` has not yet been created via
-    ///   [`Self::create_instance`].
-    pub async fn add_replica_to_instance(
-        &mut self,
-        instance_id: ComputeInstanceId,
-        replica_id: ReplicaId,
-        config: ConcreteComputeInstanceReplicaConfig,
-    ) -> Result<(), anyhow::Error> {
-        assert!(
-            self.compute.contains_key(&instance_id),
-            "call Controller::create_instance before calling add_replica_to_instance"
-        );
-
-        // Add replicas backing that instance.
-        match config.location {
-            ConcreteComputeInstanceReplicaLocation::Remote { addrs } => {
-                let mut compute_instance = self.active_compute(instance_id).unwrap();
-                compute_instance.add_replica(
-                    replica_id,
-                    addrs.into_iter().collect(),
-                    config.persisted_logs.get_sources().into_iter().collect(),
-                );
-            }
-            ConcreteComputeInstanceReplicaLocation::Managed {
-                allocation,
-                availability_zone,
-                ..
-            } => {
-                let service_name = generate_replica_service_name(instance_id, replica_id);
-
-                let service = self
-                    .compute_orchestrator
-                    .ensure_service(
-                        &service_name,
-                        ServiceConfig {
-                            image: self.computed_image.clone(),
-                            args: &|assigned| {
-                                let mut compute_opts = vec![
-                                    format!(
-                                        "--controller-listen-addr={}:{}",
-                                        assigned.listen_host, assigned.ports["controller"]
-                                    ),
-                                    format!(
-                                        "--internal-http-listen-addr={}:{}",
-                                        assigned.listen_host, assigned.ports["internal-http"]
-                                    ),
-                                    format!("--workers={}", allocation.workers),
-                                    format!("--opentelemetry-resource=instance_id={}", instance_id),
-                                    format!("--opentelemetry-resource=replica_id={}", replica_id),
-                                ];
-                                compute_opts.extend(
-                                    assigned.peers.iter().map(|(host, ports)| {
-                                        format!("{host}:{}", ports["compute"])
-                                    }),
-                                );
-                                if let Some(index) = assigned.index {
-                                    compute_opts.push(format!("--process={index}"));
-                                    compute_opts.push(format!(
-                                        "--opentelemetry-resource=replica_index={}",
-                                        index
-                                    ));
-                                }
-                                compute_opts
-                            },
-                            ports: vec![
-                                ServicePort {
-                                    name: "controller".into(),
-                                    port_hint: 2100,
-                                },
-                                ServicePort {
-                                    name: "compute".into(),
-                                    port_hint: 2102,
-                                },
-                                ServicePort {
-                                    name: "internal-http".into(),
-                                    port_hint: 6878,
-                                },
-                            ],
-                            cpu_limit: allocation.cpu_limit,
-                            memory_limit: allocation.memory_limit,
-                            scale: allocation.scale,
-                            labels: hashmap! {
-                                "replica-id".into() => replica_id.to_string(),
-                                "cluster-id".into() => instance_id.to_string(),
-                                "type".into() => "cluster".into(),
-                            },
-                            availability_zone: Some(availability_zone),
-                            // This constrains the orchestrator
-                            // (for those orchestrators that support anti-affinity, today just k8s)
-                            // to never schedule pods for different replicas of the same cluster
-                            // on the same node. Pods from the _same_ replica are fine;
-                            // pods from different clusters are also fine.
-                            //
-                            // The point is that if pods of two replicas are on the same node,
-                            // that node going down would kill both replicas, and so the replication
-                            // factor of the cluster in question is illusory.
-                            anti_affinity: Some(vec![
-                                LabelSelector {
-                                    label_name: "cluster-id".to_string(),
-                                    logic: LabelSelectionLogic::Eq {
-                                        value: instance_id.to_string(),
-                                    },
-                                },
-                                LabelSelector {
-                                    label_name: "replica-id".into(),
-                                    logic: LabelSelectionLogic::NotEq {
-                                        value: replica_id.to_string(),
-                                    },
-                                },
-                            ]),
-                        },
-                    )
-                    .await?;
-                self.active_compute(instance_id).unwrap().add_replica(
-                    replica_id,
-                    service.addresses("controller"),
-                    config.persisted_logs.get_sources().into_iter().collect(),
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Removes a replica from an instance, including its service in the
-    /// orchestrator.
-    pub async fn drop_replica(
-        &mut self,
-        instance_id: ComputeInstanceId,
-        replica_id: ReplicaId,
-        config: ConcreteComputeInstanceReplicaConfig,
-    ) -> Result<(), anyhow::Error> {
-        if let ConcreteComputeInstanceReplicaLocation::Managed { .. } = config.location {
-            let service_name = generate_replica_service_name(instance_id, replica_id);
-            self.compute_orchestrator
-                .drop_service(&service_name)
-                .await?;
-        }
-        let mut compute = self.active_compute(instance_id).unwrap();
-        compute.remove_replica(replica_id);
-        Ok(())
-    }
-
-    /// Removes an instance from the orchestrator.
-    ///
-    /// # Panics
-    /// - If the identified `instance` still has active replicas.
-    pub async fn drop_instance(
-        &mut self,
-        instance: ComputeInstanceId,
-    ) -> Result<(), anyhow::Error> {
-        if let Some(compute_state) = self.compute.remove(&instance) {
-            compute_state.drop();
-            self.compute_orchestrator
-                .drop_service(&format!("cluster-{instance}"))
-                .await?;
-        }
-        Ok(())
-    }
-
-    /// Listen for changes to compute services reported by the orchestrator.
-    pub fn watch_compute_services(&self) -> BoxStream<'static, ComputeInstanceEvent> {
-        fn translate_event(event: ServiceEvent) -> Result<ComputeInstanceEvent, anyhow::Error> {
-            let (instance_id, replica_id) = parse_replica_service_name(&event.service_id)?;
-            Ok(ComputeInstanceEvent {
-                instance_id,
-                replica_id,
-                process_id: event.process_id,
-                status: event.status,
-                time: event.time,
-            })
-        }
-
-        let stream = self
-            .compute_orchestrator
-            .watch_services()
-            .map(|event| event.and_then(translate_event))
-            .filter_map(|event| async {
-                match event {
-                    Ok(event) => Some(event),
-                    Err(error) => {
-                        tracing::error!("service watch error: {error}");
-                        None
-                    }
-                }
-            });
-
-        Box::pin(stream)
-    }
 }
 
 impl<T> Controller<T> {
-    /// Acquires an immutable handle to a controller for the storage instance.
-    #[inline]
-    pub fn storage(&self) -> &dyn StorageController<Timestamp = T> {
-        &*self.storage_controller
-    }
-
-    /// Acquires a mutable handle to a controller for the storage instance.
-    #[inline]
-    pub fn storage_mut(&mut self) -> &mut dyn StorageController<Timestamp = T> {
-        &mut *self.storage_controller
-    }
-
-    /// Acquires an immutable reference to a controller for the indicated compute instance, if it
-    /// exists.
-    #[inline]
-    pub fn compute(&self, instance: ComputeInstanceId) -> Option<&ComputeController<T>> {
-        self.compute.get(&instance)
-    }
-
-    /// Acquires a mutable reference to a controller for the indicated compute instance, if it
-    /// exists.
-    #[inline]
-    pub fn compute_mut(
-        &mut self,
-        instance: ComputeInstanceId,
-    ) -> Option<&mut ComputeController<T>> {
-        self.compute.get_mut(&instance)
-    }
-
-    /// Acquires an active controller for the indicated compute instance, if it exists.
-    #[inline]
-    pub fn active_compute(
-        &mut self,
-        instance: ComputeInstanceId,
-    ) -> Option<ActiveComputeController<T>> {
-        self.compute
-            .get_mut(&instance)
-            .map(|c| c.activate(&mut *self.storage_controller))
+    pub fn active_compute(&mut self) -> ActiveComputeController<T> {
+        self.compute.activate(&mut *self.storage)
     }
 }
 
 impl<T> Controller<T>
 where
-    T: Timestamp + Lattice + Codec64 + Copy,
+    T: Timestamp + Lattice,
     ComputeGrpcClient: ComputeClient<T>,
 {
     /// Marks the end of any initialization commands.
@@ -548,11 +141,8 @@ where
     /// and so it is important for a user to invoke this method as soon as it is comfortable.
     /// This method can be invoked immediately, at the potential expense of performance.
     pub fn initialization_complete(&mut self) {
-        self.initialized = true;
-        for (_, compute) in self.compute.iter_mut() {
-            compute.initialization_complete();
-        }
-        self.storage_mut().initialization_complete();
+        self.storage.initialization_complete();
+        self.compute.initialization_complete();
     }
 
     /// Waits until the controller is ready to process a response.
@@ -565,26 +155,14 @@ where
     /// This method is cancellation safe.
     pub async fn ready(&mut self) {
         if let Readiness::NotReady = self.readiness {
-            if self.compute.is_empty() {
-                // If there are no clients, block forever. This signals that
-                // there may be more work to do (e.g., if a compute instance
-                // is created). Calling `select_all` with an empty list of
-                // futures will panic.
-                future::pending().await
-            }
             // The underlying `ready` methods are cancellation safe, so it is
             // safe to construct this `select!`.
-            let computes = future::select_all(
-                self.compute
-                    .iter_mut()
-                    .map(|(id, compute)| Box::pin(compute.ready().map(|()| *id))),
-            );
             tokio::select! {
-                (id, _index, _remaining) = computes => {
-                    self.readiness = Readiness::Compute(id);
-                }
-                () = self.storage_controller.ready() => {
+                () = self.storage.ready() => {
                     self.readiness = Readiness::Storage;
+                }
+                id = self.compute.ready() => {
+                    self.readiness = Readiness::Compute(id);
                 }
             }
         }
@@ -601,16 +179,12 @@ where
         match mem::take(&mut self.readiness) {
             Readiness::NotReady => Ok(None),
             Readiness::Storage => {
-                self.storage_mut().process().await?;
+                self.storage.process().await?;
                 Ok(None)
             }
             Readiness::Compute(id) => {
-                let response = self
-                    .active_compute(id)
-                    .expect("reference to absent compute instance")
-                    .process()
-                    .await?;
-                Ok(response.map(|r| r.into()))
+                let response = self.active_compute().process(id).await?;
+                Ok(response.map(Into::into))
             }
         }
     }
@@ -636,14 +210,17 @@ where
             config.storaged_image,
         )
         .await;
+
+        let compute_controller = ComputeController::new(
+            config.build_info,
+            config.orchestrator.namespace("compute"),
+            config.computed_image,
+        );
+
         Self {
-            build_info: config.build_info,
-            storage_controller: Box::new(storage_controller),
-            compute_orchestrator: config.orchestrator.namespace("compute"),
-            computed_image: config.computed_image,
-            compute: BTreeMap::default(),
+            storage: Box::new(storage_controller),
+            compute: compute_controller,
             readiness: Readiness::NotReady,
-            initialized: false,
         }
     }
 }
